backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
litellm==1.50.0
pydantic==2.10.3
google-genai>=1.9.0
requests==2.31.0
pillow>=10.0.0
numpy>=1.24.0
